requirements.txt
PyQRCode==1.2.1
Pillow==7.2.0